fix: correctly handle route schema defaults with Konnect
Kong Gateway supports different schemas for different router versions.
On the other hand, Konnect can support only one schema including all
fields from 'traditional' and 'expressions' router schemas.
This may be problematic when it comes to defaults injection, because
the defaults for the 'traditiona' router schema can be wrongly injected
into the 'expressions' route configuration.

Here we make sure that only the fields that are supported for a given
router version are set in the route configuration.