fix: correct plugins config comparison
This commit fixes a bug with the way Plugins config
are compared: when unsorted nested arrays are present,
`reflect.DeepEqual` considers the equal unsorted arrays
as unequal, triggering the plugin update and showing
misleading diffs.

This commit is introducing a few helper functions to
recursively sort arrays within the Kong Plugins configs.