fix: correctly compare compressed/expanded ipv6 in targets
When syncing a target with a compressed IPv6 (for example
`[2001:db8:fd73::e]:1326`) decK can successfully create it
but it fails to detect its existence when pulling the config
from the API because the CP expands the address (for example to
`[2001:0db8:fd73:0000:0000:0000:0000:000e]:1326`) causing
decK to attempt to recreate the target and failing because the
target already exists.

This commit fixes this by using some helpers to expand the
ipv6 before trying to sync/dump.