test: Valid expression routes building (#117)
* fix: Fix yaml files rendered for expression routes

At the moment, when we run deck file render, all routes get
`regex_priority` field. This field is not applicable in the case
of expression routes. A rendered file from deck containing
expression routes, thus failed to sync directly.

This change ensures that expression routes do not get the
fields of traditional routes.

Fix #1250

* test: Valid expression routes building

Test is added for changes to builder.go which ensure that
expression routes do not contain regex_priority field.
Separated the tests for traditional and expression routes
for clarity.

This test goes along with fix #1250.

* test: Added test for isKonnect: false in builder

Test updated to check for route building in case
isKonnect is false ie on command runs like
deck file render.