chore: update go-kong to v0.57.1
The new go-kong version ensures that shorthand_fields are
filled with defaults or nested config values. If the shorthand_fields
are set, the nested fields are backfilled to keep both in sync.
This ensures that deck diff or deck sync commands do not show
unnecessary updates in plugin configs.

Fix for #1251 from deck issues.