fix(diff+file): avoid filling defaults in config for kong (#133)
* fix(diff+file): avoid filling defaults in config for kong

the previous logic was filling defaults in the configuration that was
passed to Kong. This was  problematic, especially where nils were
populated as defaults, e.g. if a shorthand_field was passed with some
value and the corresponding new field is auto-populated as `nil` by decK
, the auto-populated nil value would take precedence in Kong thus causing
the shorthand_field to be ignored
(https://konghq.atlassian.net/browse/KAG-5157).

This change applies the default values only to configurations used for
diff, the original configuration is always passed to Kong as is.

* Test/avoid filling defaults in config for kong (#135)

* add tests for plugin filling default values

---------

Co-authored-by: samugi <samuele@konghq.com>

---------

Co-authored-by: Tao Yi <richardyi0110@hotmail.com>