fix: updating Konnect auth logic to properly handle geo rewrites (#146)
As Konnect's authorization service is not deployed across all regions,
the global endpoint is used for certain API calls. When the organization
info is fetched, the hostname within the provided base URL is then
rewritten to the relevant global endpoint.

This rewrite logic was unable to handle any number of geos, relying on
hard coded logic instead for specific geos. This logic has been updated
to allow callers of this library to talk to multiple geos; this includes
new geos that may not exist as of this writing.