feat: ignoring konnect-managed entities with deck
As of now, deck supports ignoring plugins
that are konnect-managed. We are adding further
support to ignore any entity that may have the
tag "konnect-managed" added to it. Konnect
adds these tags to some entities and they
are treated as immutable. Thus, we wish to
ignore any deck operations on them.

With this change, a deck dump would not
include the "konnect-managed" entities.
deck sync/diff would also ignore these entities.
For: https://github.com/Kong/deck/issues/1442