fix: ensuring consumer-group scoped plugins show tags (#151)
* fix: ensuring consumer-group scoped plugins show tags

On running `deck file render` command, tags
were coming up for global and service/route
scoped plugins, but not for consumer-group
scoped ones. This was because the underlying
struct that were using for writer did not have
`Tags` as a field. This change was added in the
underlying type in go-kong.

This change upgrades the go-kong version to
v0.61.0 and adds `Tags` field in all missing
places while building consumer-group scoped
plugins or writing them to file.
For: https://konghq.atlassian.net/browse/FTI-6292

* chore: added test-case for cg scoped plugins with tags