Add support for foreign key references in partial sync
The b.intermediate collection is populated when entities
are discovered in targetState, allowing for in-file foreign
key references.

The new `deck gateway apply` command allows you to sync
partial configuration, and so must populate b.intermediate
from the currentState.

This results in duplicate entity errors, and so an
AddIgnoringDuplicates method has been added to entities
that may be a foreign key