fix: returning a validation error in case of wrong configuration. (#157)
* fix: returning a validation error in case of wrong configuration.

As of now, while using deck to create a
route-scoped plugin with a `service` field
defined, the sync goes through the first
time but errors out subsequently. Under a
route, we do not want the user to add the
`service` field for a plugin. Thus, we are
checking for this specific scenario and
erroring out, instead of letting the sync
action go through.

For: https://github.com/Kong/deck/issues/1353

* fix: added checks for all conflicting nested configs in plugins.

A foreign key nested under a plugin of a different
scope would error out. This would make sure that a
sync does not go through when wrong configurations
are passed via deck. An example would be adding a
`service` field in a plugin, nested under a route.
This change ensures more such issues do not come up.

* test: added integration test for validations related to scoped plugins

* chore: removed redundant code lines

* chore: removed stray code lines from test

* test: fixed tests and added more for kong <3.0

* chore: addressed PR comments