fix: fixing consumer-group related dump config for deck. (#159)
* fix: fixing consumer-group related dump config for deck.

With changes added in go-kong, to skip
listing consumers while doing a GET call
on a consumer_group, dump config with gw
3.9+ starting showing issues. Associations
between a consumer and consumer-group were
missing. Thus, we are pushing the no listing
consumers feature behind a flag for now,
applicable only for dump command as a
starting point.

This change adds a new config that can be
set via a deck command flag. If set to true,
it will not list consumers-consumer group
associations in the dump file.

* chore: updated go-kong to v0.63.0