fix: fetching plugin-relations from intermediate state too. At the moment, we are fetching plugin-relations only from currentState. In scenarios where the currentState consists of a global plugin X, syncing a different state with plugin X associated with service S leads to a failure - "entity already exists". However, both of these plugin instances are unique and should not result in duplication errors.
This happened because we
were skipping the check
for intermediate state.
The new service created with
second sync would be a part
of intermediate state, not
current state. So, if
we check only currentState, we
miss the serviceID and thus
ingesting a global plugin again,
which led to the duplication
error.